Make app a package and move tests out
