+++ app/__init__.py
@@ -1,16 +1,17 @@
 from os import path as os_path
 
 from flask import Flask, send_file, send_from_directory
-from resources.magic_castle_api import MagicCastleAPI
-from resources.progress_api import ProgressAPI
-from resources.available_resources_api import AvailableResourcesApi
-from resources.user_api import UserAPI
 from flask_cors import CORS
-from database.schema_manager import SchemaManager
-from database.database_manager import DatabaseManager
-from models.configuration import config
 
-from models.constants import DIST_PATH
+from . database.schema_manager import SchemaManager
+from . database.database_manager import DatabaseManager
+from . models.configuration import config
+from . models.constants import DIST_PATH
+from . resources.magic_castle_api import MagicCastleAPI
+from . resources.progress_api import ProgressAPI
+from . resources.available_resources_api import AvailableResourcesApi
+from . resources.user_api import UserAPI
+
 
 # Update the database schema to the latest version
 with DatabaseManager.connect() as database_connection:
@@ -99,11 +100,3 @@ def catch_all(path):
     response.headers["Pragma"] = "no-cache"
     response.headers["Expires"] = "0"
     return response
-
-
-if __name__ == "__main__":
-    app.run(
-        host="0.0.0.0",
-        port=config["port"],
-        debug=config["debug"]
-    )
