+++ app/database/schema_manager.py
@@ -1,8 +1,9 @@
 import sqlite3
+
 from os import listdir, path
 from os.path import isfile
-from models.constants import SCHEMA_MIGRATIONS_DIRECTORY
 
+from .. models.constants import SCHEMA_MIGRATIONS_DIRECTORY
 
 class SchemaManager:
     """
