+++ app/models/magic_castle/magic_castle_configuration_schema.py
@@ -1,7 +1,10 @@
-from marshmallow import Schema, fields, validate
-from models.constants import INSTANCE_CATEGORIES
-from models.cloud.dns_manager import DnsManager
 import re
+
+from marshmallow import Schema, fields, validate
+
+from .. constants import INSTANCE_CATEGORIES
+from .. cloud.dns_manager import DnsManager
+
 
 def validate_cluster_name(cluster_name):
     # Must follow RFC 1035's subdomain naming rules: https://tools.ietf.org/html/rfc1035#section-2.3.1
