+++ app/models/user/authenticated_user.py
@@ -1,9 +1,9 @@
 import json
 
-from models.user.user import User
-from models.magic_castle.magic_castle import MagicCastle
-from models.configuration import config
-from models.constants import DEFAULT_CLOUD
+from . user import User
+from .. magic_castle.magic_castle import MagicCastle
+from .. configuration import config
+from .. constants import DEFAULT_CLOUD
 
 
 class AuthenticatedUser(User):
