+++ app/models/user/user.py
@@ -1,9 +1,9 @@
 import sqlite3
 
 from typing import List
-from models.magic_castle.magic_castle import MagicCastle
-from models.constants import DEFAULT_CLOUD
 
+from .. magic_castle.magic_castle import MagicCastle
+from .. constants import DEFAULT_CLOUD
 
 class User:
     def __init__(self, database_connection: sqlite3.Connection):
