+++ app/resources/magic_castle_api.py
@@ -1,7 +1,7 @@
 from flask import request
-from resources.api_view import ApiView
-from exceptions.invalid_usage_exception import InvalidUsageException
-from models.user.user import User
+from . api_view import ApiView
+from .. exceptions.invalid_usage_exception import InvalidUsageException
+from .. models.user.user import User
 
 
 class MagicCastleAPI(ApiView):
