+++ app/resources/progress_api.py
@@ -1,7 +1,7 @@
-from resources.api_view import ApiView
-from exceptions.invalid_usage_exception import InvalidUsageException
-from models.magic_castle.cluster_status_code import ClusterStatusCode
-from models.user.user import User
+from . api_view import ApiView
+from .. exceptions.invalid_usage_exception import InvalidUsageException
+from .. models.magic_castle.cluster_status_code import ClusterStatusCode
+from .. models.user.user import User
 
 
 class ProgressAPI(ApiView):
