+++ tests/mock-clusters/missingfloatingips.c3.ca/main.tf.json
