+++ tests/mock-clusters/missingfloatingips.c3.ca/terraform_plan.json
