+++ tests/mock-clusters/missingnodes.sub.example.com/terraform.tfstate
