+++ tests/mock-clusters/valid1.calculquebec.cloud/terraform_apply.log
