+++ tests/mocks/openstack/openstack_connection_mock.py
