+++ tests/unit/terraform/test_terraform_plan_parser.py
@@ -1,9 +1,12 @@
-from models.terraform.terraform_plan_parser import TerraformPlanParser
 import pytest
+
 import tests
 import json
-from tests.test_helpers import *  # noqa
-from tests.mocks.configuration.config_mock import config_auth_none_mock  # noqa;
+
+from app.models.terraform.terraform_plan_parser import TerraformPlanParser
+
+from ... test_helpers import *  # noqa
+from ... mocks.configuration.config_mock import config_auth_none_mock  # noqa;
 
 
 def load_plan(hostname):
